Add user paging,login,register demo

+++ core/dto/dto.go
@@ -1,9 +1,9 @@
 package dto
 
 type Pagination[T any] struct {
-	Items       []T `json:"items"`
-	TotalItems  int `json:"total_items"`
-	TotalPages  int `json:"total_pages"`
-	CurrentPage int `json:"current_page"`
-	PageSize    int `json:"page_size"`
+	Items       []T   `json:"items"`
+	TotalItems  int64 `json:"total_items"`
+	TotalPages  int64 `json:"total_pages"`
+	CurrentPage int   `json:"current_page"`
+	PageSize    int   `json:"page_size"`
 }
