+++ core/entity/entity.go
@@ -2,8 +2,8 @@ package entity
 
 type Pagination[T any] struct {
 	Items       []T
-	TotalItems  int
-	TotalPages  int
+	TotalItems  int64
+	TotalPages  int64
 	CurrentPage int
 	PageSize    int
 }
